refactor: route group을 통해 CommonNavbar가 있는 것을 통합한다
- common/nabvar안에 Buton, Navbar, NavbarCell, UserInfo의 component
  postfix를 제거하였습니다.
- NavbarOperation을 추가하여 client side rendering되는 것을 빼냈습니다.
- navbar를 사용하는 admin, applicant, interview, kanban을 빼내었고 이를
  WithNavbar로 통합하여 layout을 그립니다.
