+++ frontend/app/(WithNavbar)/kanban/[generation]/detail/page.tsx
