+++ frontend/app/(WithNavbar)/layout.tsx
@@ -1,0 +1,27 @@
+import CommonNavbar from "@/components/common/navbar/Navbar";
+import Validate from "@/components/user/Validate.component";
+import { PropsWithChildren } from "react";
+import { headers } from "next/headers";
+
+interface WithNavbarLayout extends PropsWithChildren {}
+
+const ApplicantPage = ({ children }: WithNavbarLayout) => {
+  const headersList = headers();
+  const header_url = headersList.get("x-url") || "";
+  const [_, __, ___, currentPath, generation] = header_url.split("/");
+  const isShort = currentPath === "kanban" ? true : false;
+
+  return (
+    <div className="px-24 min-w-[1280px] flex p-12">
+      <Validate />
+      <CommonNavbar
+        generation={generation}
+        currentPath={currentPath}
+        isShort={isShort}
+      />
+      {children}
+    </div>
+  );
+};
+
+export default ApplicantPage;
